Calculate/display ping on both ends.
It would be more efficient to still only send pings from one side, and
include the last measurement in each one, so that the other peer doesn't
need to take its own measurements. However, doing the same thing on both
ends makes the code simpler, and 2x(one ping per second) is still not a
lot.