Delete user data when they leave the page.
A couple caveats:
1. onbeforeunload apparently doesn't work if the user hasn't interacted
   with the page, so it's possible that a user just opens and closes the
   page and their data isn't deleted.
2. I'm sure there are some circumstances where the event isn't
   triggered. (ie if the browser crashes.)
3. The firestore API doesn't allow recursively deleting subcollections
   of a document from the web, so child data (such as ICE candidates)
   are still left around.

I'm ok with 1 and 2 for now because this is better than nothing. I could
re-add the "get connected" button to force user interaction, but I don't
think the increased annoyance to users and extra code path is worth it
at the moment. Once I add react or something it'll be a lot cleaner to
do something like that.

I'm ok with 3 because even if the child data is hanging around it's
still inaccessible since there is no longer a parent referencing it. Somebody
just spamming room IDs would not stumble onto it.