Quadruple the ping rate and remove the y axis label.
I'll re-add the label once I can figure out a way to make it meaningful.
In the meantime, people will get the idea.

Quadrupling the ping rate makes the graph look a lot smoother. More
importantly, it fixes the issue with the line sometimes being visible to
the left of the y-axis. Since the graph moves 1/4 the distance it
normally would when a new point is added, it no longer moves far enough
to the left for old points to be visible before they are deleted.